extract choices listing selector into component
