+++ tutor/resources/styles/components/choices-listing.less
@@ -1,0 +1,77 @@
+.make-wizard-choice-active() {
+  background-color: fade(@tutor-neutral-light, 60%);
+  border-color: @tutor-neutral-light;
+  color: inherit;
+  position: relative;
+}
+
+.choices-listing {
+  @choice-height: 70px;
+
+  max-height: @choice-height * 3.5;
+  overflow-y: auto;
+  border-top: 1px solid @tutor-neutral-light;
+  border-bottom: 1px solid @tutor-neutral-light;
+
+  .list-group-item:first-child {
+    border-top: none;
+  }
+  .list-group-item:last-child {
+    border-bottom: none;
+  }
+
+  .choice {
+
+    padding: 8px;
+    display: flex;
+    align-content: center;
+    height: @choice-height;
+    position: relative;
+    font-size: 1.6rem;
+    font-weight: 800;
+    line-height: 1em;
+    flex-direction: row;
+    align-items: center;
+    cursor: pointer;
+    &:focus {
+      outline: none;
+    }
+    &::before {
+      content: '';
+      background-size: contain;
+      background-repeat: no-repeat;
+      height: 2.5rem;
+      margin-right: 1rem;
+      display: block;
+      margin-top: 0;
+    }
+    &::after {
+      content: '';
+      position: absolute;
+      background-repeat: no-repeat;
+      background-position: center center;
+      display: block;
+    }
+
+    &.active {
+      .make-wizard-choice-active();
+      &:hover,
+      &:focus {
+        .make-wizard-choice-active();
+      }
+      &::after {
+        right: 16px;
+        top: calc(50% - 4px);
+        .fa-icon();
+        content: @fa-var-check;
+        color: @tutor-secondary;
+        margin-bottom: 0;
+        font-size: 20px;
+        margin-top: -11px;
+        background-color: @tutor-white;
+        border-radius: 50%;
+        padding: 4px;
+      }
+    }
+  }
+}
