+++ tutor/resources/styles/tutor.less
@@ -85,6 +85,8 @@
 @import './components/info-page';
 @import './components/terms-modal';
 @import './components/payments';
+@import './components/choices-listing';
+
 
 body[data-js-error]::before {
   content: attr(data-js-error);
